replace std::make_unique patch with boost::make_unique
On my system libear failed to build because no header defining
__cpp_lib_make_unique was included, so std::make_unique was defined
twice.

Looking further, it's not clear that these macros were actually required
in C++14; they seem to be a later feature, and anyway some versions of
MSVC define std::make_unique but don't define this macro, so bad things
could happen there.

The behaviour is undefined when adding things to the std namespace, so
we shouldn't do that anyway. Just boost until we decide to switch to
C++14.