allow relative imports of generated files
For each installed generated file there is a file in /include where it
is referenced. During the build, this includes a file from the build
dir, which is on the include path.  When it's installed, this is a
relative include instead.

For example, in `include/ear/export.hpp`:

    #include "generated/export.hpp"

When building in `build`, this references `build/generated/export.hpp`,
because `build` is on the include path.

When installed, this would get installed as e.g.
`/usr/include/ear/export.hpp`, and always references
`/usr/include/ear/generated/export.hpp` no matter what the include path
is, because it acts as a relative path.