implement extent panner using a hierarchical triangular mesh
This is an alternative to the standard extent panner, which improves
performance but gives the same results.

The sphere is split into triangles, and the panning positions are
distributed between the triangles. The total panning gains for the whole
sphere is the sum of the panning gains for each triangle, which can be
calculated efficiently in most cases:

- triangles completely inside the extent shape have weight==1 for all
  points, so the total can be pre-computed.

- triangles completely outside the extent shape have weight==0 for all
  points, to have total panning values of 0.

- the weight for all points in triangles which cross between the inside
  and outside have to be calculated and summed as in the existing panner

The first two cases account for the majority of triangles, so as long as
the test for whether a triangle is inside or outside the extent is
reasonably efficient, this is beneficial.

The inside/outside testing is performed by WeightingFunctionEdges, which
uses the GreatCircleArcIntersection class.

SpreadingPannerHTM extends SpreadingPannerBase, and contains the above
logic for summing point weights using a triangulation. This uses the HTM
class, which represents a recursive sub-division of a sphere into
triangles.

This represents a substantial increase in complexity, diverging
significantly from the standard (in implementation, not results), so has
been made optional.