Use cmake rather than metaprogramme to embed text resource files
When cross compiling, the resource_embedder tool currently fails. This
is because it is built with the target architecture but needs to be run
under the host architecture.

It would be possible to treat the resource embedder as an external
project (either via ExternalProject_Add or a custom target that calls
cmake) but that feels complicated and potentially error prone if some of
the cross compile setup is done via environment variables.

This commit replaces the resource embedder executable with a cmake
script to generate a similar file, as cmake is always run on the host
platform it resolves the issues.

The generated implementation differs a little from that generated via
the resource embedder executable:

* The function to retrieve files is no longer inline and is nested in a
  namespace to reduce the chance of violating ODR.

* The individual character arrays that store each file are in an
  anonymous namespace in a cpp file rather than static vars in a header.

* No attempt is made to correctly format the generated files as they are
  not intended to be user readable/editable.